# docs/swagger.json
{"openapi":"3.0.0","info":{"title":"CLARK Gateway","version":"2.26.6","description":"The API gateway for all of the CLARK microservices","license":{"name":"ISC","url":"https://www.isc.org/licenses/"},"contact":{"name":"SecurEd Inc","url":"http://secured.team/","email":"skaza@towson.edu"},"x-logo":{"url":"http://secured.team/assets/images/logo.png","altText":"CLARK Logo"},"x-tagGroups":[{"name":"Library Service","tags":["Library Service"]},{"name":"Notification Service","tags":["Notification Service"]},{"name":"Feature Service","tags":["Feature Service"]},{"name":"Rating Service","tags":["Rating Service"]},{"name":"Outcome Service","tags":["Outcome Service"]},{"name":"Learning Object Service","tags":["Learning Object Service"]}]},"servers":[{"url":"http://localhost:3001","description":"Development"},{"url":"https://api-gateway.clark.center","description":"Production"}],"paths":{"/featured/learning-objects":{"get":{"description":"Gets a list of featured objects","tags":["Feature Service"],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LearningObject"}}}}}}},"patch":{"description":"Updates the list of featured objects","tags":["Feature Service"],"requestBody":{"description":"The cuid, version, and author's username","required":true,"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LearningObject"}}}}},"responses":{"201":{"description":"CREATED","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string","example":"Learning Objects Updated"}}}}}},"401":{"description":"UNAUTHENTICATED - User not logged in"},"403":{"description":"UNAUTHORIZED - User does not have privilege to update featured objects"}}}},"/users/{username}/learning-objects/{learningObjectId}/outcomes/{outcomeId}/mappings/{mappingId}":{"delete":{"description":"Removes a mapping from a object's outcome","tags":["Guideline Service"],"parameters":[{"in":"path","name":"username","schema":{"type":"string"},"required":true,"description":"The username of the author for the learning object"},{"in":"path","name":"learningObjectId","schema":{"type":"string"},"required":true,"description":"The id of the learning object (not cuid)"},{"in":"path","name":"outcomeId","schema":{"type":"string"},"required":true,"description":"The id of the outcome"},{"in":"path","name":"mappingId","schema":{"type":"string"},"required":true,"description":"The id of the mapping"}],"responses":{"204":{"description":"NO CONTENT"},"401":{"description":"UNAUTHENTICATED - User not logged in"},"403":{"description":"UNAUTHORIZED - Another user is trying to access a author's outcome on a object"},"404":{"description":"NOT FOUND - Object, mapping, outcome, or user not found"}}}},"/outcomes/stats":{"get":{"description":"Gets the stats of outcomes","tags":["Guideline Service"],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","properties":{"apply":{"type":"integer","required":true,"example":10},"evaluate":{"type":"integer","required":true,"example":20},"remember":{"type":"integer","required":true,"example":30}}}}}}}}},"/users/{userId}/learning-objects/{cuid}/changelog":{"post":{"description":"Creates a changelog","tags":["Learning Object Service"],"parameters":[{"in":"path","name":"userId","schema":{"type":"string"},"required":true,"description":"The username of the author"},{"in":"path","name":"cuid","schema":{"type":"string"},"required":true,"description":"The cuid of the learning object"}],"requestBody":{"description":"The cuid, version, and author's username","required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"changelogText":{"type":"string","description":"This is the text of the changelog","required":true,"example":"This is an example"}}}}}},"responses":{"200":{"description":"OK"},"401":{"description":"UNAUTHENTICATED - User not logged in"},"403":{"description":"UNAUTHORIZED - User is not a admin or editor"},"404":{"description":"NOT FOUND - Learning object or user not found"}}}},"/users/{userId}/learning-objects/{cuid}/changelogs":{"get":{"description":"Gets a changelog for a object","tags":["Learning Object Service"],"parameters":[{"in":"path","name":"userId","schema":{"type":"string"},"required":true,"description":"The username of the author"},{"in":"path","name":"cuid","schema":{"type":"string"},"required":true,"description":"The cuid of the learning object"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/Changelog"}}}}}}},"/learning-objects/metrics":{"get":{"description":"Gets the learning object metrics for CLARK","tags":["Library Service"],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Metric"}}}},"404":{"description":"NOT FOUND - No metrics were found"}}}},"/users/{username}/learning-objects/{cuid}/metrics":{"get":{"description":"Gets the metrics for a learning object","tags":["Library Service"],"parameters":[{"in":"path","name":"username","schema":{"type":"string"},"required":true,"description":"The author's username"},{"in":"path","name":"cuid","schema":{"type":"string"},"required":true,"description":"The cuid of the learning object"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","properties":{"saves":{"type":"integer","description":"The number of saves for a particular object","example":200},"downloads":{"type":"integer","description":"The number of downloads for a particular object","example":100}}}}}},"404":{"description":"NOT FOUND - No metrics were found for the specific username and cuid"}}}},"/{username}/library/learning-objects":{"get":{"description":"Gets a user's library","tags":["Library Service"],"parameters":[{"in":"path","name":"username","schema":{"type":"string"},"required":true,"description":"The username of the user's library"},{"in":"query","name":"page","schema":{"type":"integer"},"required":true,"description":"The current page of the library"},{"in":"query","name":"limit","schema":{"type":"integer"},"required":true,"description":"The number of objects to return"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","properties":{"userLibraryItems":{"type":"array","items":{"$ref":"#/components/schemas/LibraryItem"},"required":true},"lastPage":{"type":"integer","required":false,"example":2}}}}}},"401":{"description":"UNAUTHENTICATED - User not logged in"},"403":{"description":"UNAUTHORIZED - User is trying to access another user's library"},"404":{"description":"NOT FOUND - User not found"}}},"post":{"description":"Adds a learning object to the user's library","tags":["Library Service"],"parameters":[{"in":"path","name":"username","schema":{"type":"string"},"required":true,"description":"The username of the user's library"}],"requestBody":{"description":"The cuid, version, and author's username","required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"cuid":{"type":"string","description":"The cuid of the learning object to save","required":true,"example":"00000000-1111-2222-3333-444444444444"},"version":{"type":"integer","description":"The version of the learning object","required":true,"example":1},"authorUsername":{"type":"string","description":"The username of the object's author","required":true,"example":"jdoe1"}}}}}},"responses":{"201":{"description":"CREATED - Successfully added to library"},"401":{"description":"UNAUTHENTICATED - User not logged in"},"403":{"description":"UNAUTHORIZED - User is trying to access and modify another user's library"},"404":{"description":"NOT FOUND - User or learning object not found"}}}},"/{username}/library/learning-objects/{cuid}":{"delete":{"description":"Deletes an item from the user's library","tags":["Library Service"],"parameters":[{"in":"path","name":"username","schema":{"type":"string"},"required":true,"description":"The username of the user's library"},{"in":"path","name":"cuid","schema":{"type":"string"},"required":true,"description":"The cuid of the learning object to remove"}],"responses":{"204":{"description":"NO CONTENT - Successfully deleted from library"},"401":{"description":"UNAUTHENTICATED - User not logged in"},"403":{"description":"UNAUTHORIZED - User is trying to access and modify another user's library"},"404":{"description":"NOT FOUND - User or learning object not found"}}}},"/users/{username}/notifications":{"get":{"description":"Gets a user's update notifications","tags":["Notification Service"],"parameters":[{"in":"path","name":"username","schema":{"type":"string"},"required":true,"description":"The username of the user"},{"in":"query","name":"page","schema":{"type":"number"},"required":true,"description":"The current page"},{"in":"query","name":"limit","schema":{"type":"number"},"required":true,"description":"The notification limit to send back in the request"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","properties":{"lastPage":{"type":"number","example":1,"required":true},"notifications":{"type":"array","items":{"$ref":"#/components/schemas/Notification"},"required":true}}}}}},"400":{"description":"BAD CONTENT - Page and limit query parameters must be greater then zero"},"401":{"description":"UNAUTHENTICATED - User not logged in"},"403":{"description":"UNAUTHORIZED - User is trying to access another user's notifications"},"404":{"description":"NOT FOUND - User not found"}}}},"/users/{username}/notifications/{id}":{"delete":{"description":"Deletes a notification for a user","tags":["Notification Service"],"parameters":[{"in":"path","name":"username","schema":{"type":"string"},"required":true,"description":"The username of the user"},{"in":"path","name":"id","schema":{"type":"string"},"required":true,"description":"The id of the notification"}],"responses":{"204":{"description":"NO CONTENT"},"401":{"description":"UNAUTHENTICATED - User not logged in"},"403":{"description":"UNAUTHORIZED - User is trying to delete another user's notifications"},"404":{"description":"NOT FOUND - User or notification not found"}}}},"/users/{username}/learning-objects/{CUID}/version/{version}/ratings":{"get":{"description":"Gets the ratings for a learning object","tags":["Rating Service"],"parameters":[{"in":"path","name":"username","schema":{"type":"string"},"required":true,"description":"The username of the object's author"},{"in":"path","name":"CUID","schema":{"type":"string"},"required":true,"description":"The cuid of the object"},{"in":"path","name":"version","schema":{"type":"number"},"required":true,"description":"The version number of the object"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","properties":{"_id":{"type":"string","description":"The cuid of the object","example":"00000000-1111-2222-3333-444444444444"},"avgValue":{"type":"number","description":"The average rating value","example":5},"ratings":{"type":"array","description":"The array of ratings left on the object","items":{"$ref":"#/components/schemas/Rating"}}}}}}},"404":{"description":"NOT FOUND - Learning object not found"}}},"post":{"description":"Creates a rating on a object","tags":["Rating Service"],"parameters":[{"in":"path","name":"username","schema":{"type":"string"},"required":true,"description":"The username of the object's author"},{"in":"path","name":"CUID","schema":{"type":"string"},"required":true,"description":"The cuid of the object"},{"in":"path","name":"version","schema":{"type":"number"},"required":true,"description":"The version number of the object"}],"requestBody":{"description":"The rating information to create","required":true,"content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/Rating"}}}},"responses":{"204":{"description":"NO CONTENT"},"401":{"description":"UNAUTHENTICATED - User not logged in"},"403":{"description":"UNAUTHORIZED - Object attempting to rate is in review (not released)"},"404":{"description":"NOT FOUND - Learning object was not found"}}}},"/ratings/{ratingId}":{"get":{"description":"Gets a specific rating by id","tags":["Rating Service"],"parameters":[{"in":"path","name":"ratingId","schema":{"type":"string"},"required":true,"description":"The rating id"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/Rating"}}}},"404":{"description":"NOT FOUND - Rating was not found"}}}},"/users/{username}/learning-objects/{CUID}/version/{version}/ratings/{ratingID}":{"patch":{"description":"Updates a rating on a object","tags":["Rating Service"],"parameters":[{"in":"path","name":"username","schema":{"type":"string"},"required":true,"description":"The username of the object's author"},{"in":"path","name":"CUID","schema":{"type":"string"},"required":true,"description":"The cuid of the object"},{"in":"path","name":"version","schema":{"type":"number"},"required":true,"description":"The version number of the object"},{"in":"path","name":"ratingID","schema":{"type":"string"},"required":true,"description":"The id of the rating to update"}],"requestBody":{"description":"The rating information to update","required":true,"content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/Rating"}}}},"responses":{"200":{"description":"OK"},"401":{"description":"UNAUTHENTICATED - User not logged in"},"403":{"description":"UNAUTHORIZED - User is trying to access and modify another user's rating"},"404":{"description":"NOT FOUND - Rating or object was not found"}}},"delete":{"description":"Deletes a rating from a learning object","tags":["Rating Service"],"parameters":[{"in":"path","name":"username","schema":{"type":"string"},"required":true,"description":"The username of the object's author"},{"in":"path","name":"CUID","schema":{"type":"string"},"required":true,"description":"The cuid of the object"},{"in":"path","name":"version","schema":{"type":"number"},"required":true,"description":"The version number of the object"},{"in":"path","name":"ratingID","schema":{"type":"string"},"required":true,"description":"The id of the rating to delete"}],"responses":{"200":{"description":"OK"},"401":{"description":"UNAUTHENTICATED - User not logged in"},"403":{"description":"UNAUTHORIZED - User is trying to access and modify another user's rating"},"404":{"description":"NOT FOUND - Rating or object was not found"}}}},"/users/{username}/learning-objects/{CUID}/version/{version}/ratings/{ratingID}/flags":{"post":{"description":"Flags a rating to be reviewed","tags":["Rating Service"],"parameters":[{"in":"path","name":"username","schema":{"type":"string"},"required":true,"description":"The username of the object's author"},{"in":"path","name":"CUID","schema":{"type":"string"},"required":true,"description":"The cuid of the object"},{"in":"path","name":"version","schema":{"type":"number"},"required":true,"description":"The version number of the object"},{"in":"path","name":"ratingID","schema":{"type":"string"},"required":true,"description":"The id of the rating"}],"requestBody":{"description":"The flag information to create","required":true,"content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/Flag"}}}},"responses":{"204":{"description":"NO CONTENT"},"401":{"description":"UNAUTHENTICATED - User not logged in"},"403":{"description":"UNAUTHORIZED - Author is attempting to flag a rating"},"404":{"description":"NOT FOUND - Learning object or rating was not found"}}}},"/users/{username}/learning-objects/{CUID}/version/{version}/ratings/{ratingID}/responses":{"post":{"description":"Creates a response to a rating","tags":["Rating Service"],"parameters":[{"in":"path","name":"username","schema":{"type":"string"},"required":true,"description":"The username of the object's author"},{"in":"path","name":"CUID","schema":{"type":"string"},"required":true,"description":"The cuid of the object"},{"in":"path","name":"version","schema":{"type":"number"},"required":true,"description":"The version number of the object"},{"in":"path","name":"ratingID","schema":{"type":"string"},"required":true,"description":"The id of the rating"}],"requestBody":{"description":"The flag information to create","required":true,"content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/Response"}}}},"responses":{"200":{"description":"OK"},"401":{"description":"UNAUTHENTICATED - User not logged in"},"403":{"description":"UNAUTHORIZED - Is not author or contributor"},"404":{"description":"NOT FOUND - Learning object or rating was not found"}}}},"/users/{username}/learning-objects/{CUID}/version/{version}/ratings/{ratingID}/responses/{responseID}":{"delete":{"description":"Deletes a response","tags":["Rating Service"],"parameters":[{"in":"path","name":"username","schema":{"type":"string"},"required":true,"description":"The username of the object's author"},{"in":"path","name":"CUID","schema":{"type":"string"},"required":true,"description":"The cuid of the object"},{"in":"path","name":"version","schema":{"type":"number"},"required":true,"description":"The version number of the object"},{"in":"path","name":"ratingID","schema":{"type":"string"},"required":true,"description":"The id of the rating"},{"in":"path","name":"responseID","schema":{"type":"string"},"required":true,"description":"The id of the response"}],"responses":{"200":{"description":"OK"},"401":{"description":"UNAUTHENTICATED - User not logged in"},"403":{"description":"UNAUTHORIZED - Is not the creator of the response"},"404":{"description":"NOT FOUND - Learning object, rating, or response was not found"}}},"patch":{"description":"Updates a response","tags":["Rating Service"],"parameters":[{"in":"path","name":"username","schema":{"type":"string"},"required":true,"description":"The username of the object's author"},{"in":"path","name":"CUID","schema":{"type":"string"},"required":true,"description":"The cuid of the object"},{"in":"path","name":"version","schema":{"type":"number"},"required":true,"description":"The version number of the object"},{"in":"path","name":"ratingID","schema":{"type":"string"},"required":true,"description":"The id of the rating"},{"in":"path","name":"responseID","schema":{"type":"string"},"required":true,"description":"The id of the response"}],"requestBody":{"description":"The flag information to create","required":true,"content":{"application/json":{"schema":{"type":"object","$ref":"#/components/schemas/Response"}}}},"responses":{"200":{"description":"OK"},"401":{"description":"UNAUTHENTICATED - User not logged in"},"403":{"description":"UNAUTHORIZED - Is not the creator of the response"},"404":{"description":"NOT FOUND - Learning object, rating, or response was not found"}}}}},"components":{"schemas":{"LearningObject":{"properties":{"id":{"type":"string","description":"The id of the learning object","example":0,"required":true},"cuid":{"type":"string","description":"The cuid of the learning object","example":"00000000-1111-2222-3333-444444444444","required":true},"author":{"type":"object","description":"A partial user object of the learning object's author","$ref":"#/components/schemas/User","required":true},"collection":{"type":"string","description":"The collection the learning object belongs to","example":"nccp","required":true},"contributors":{"type":"array","description":"A optional array of extra contributors to the object","items":{"$ref":"#/components/schemas/User"},"required":true},"date":{"type":"string","description":"The date the learning object was created","example":0,"required":true},"description":{"type":"string","description":"The description of the learning object","example":"This is a description","required":true},"length":{"type":"string","description":"The length of the learning object","example":"nanomodule","required":true},"name":{"type":"string","description":"The name of the learning object","example":"This is a learning object name","required":true},"status":{"type":"string","description":"The current status of the learning object","example":"released","required":true},"resourceUris":{"type":"object","description":"The extra resource uris of the learning object","required":false,"$ref":"#/components/schemas/LearningObjectResourceUris"},"revisionUri":{"type":"string","description":"The new revision uri of the learning object","required":true},"version":{"type":"integer","description":"The current version of the learning object","example":1,"required":true},"levels":{"type":"array","description":"An array of appropriate academic levels to the object","items":{"type":"string","example":"undergraduate"},"required":true}}},"User":{"properties":{"id":{"type":"string","description":"The id of the user","example":0,"required":true},"username":{"type":"string","description":"The username of the user","example":"jdoe1","required":true},"name":{"type":"string","description":"The name of the user","example":"John Doe","required":true},"email":{"type":"string","description":"The email of the user","example":"jdoe1@gmail.com","required":true},"emailVerified":{"type":"boolean","description":"A boolean representing if the user's email has been verified","example":true,"required":true},"organization":{"type":"string","description":"The organization of the user","example":"Towson University","required":true},"bio":{"type":"string","description":"A short biography of the user","example":"This is a bio","required":true},"createdAt":{"type":"string","description":"The date when the user was created","example":0,"required":true}}},"LearningObjectResourceUris":{"properties":{"outcomes":{"type":"string","description":"The uri for the outcomes of a learning object","example":"https://api-gateway.clark.center/users/jdoe1/learning-objects/000000000000000000000000/outcomes","required":false},"children":{"type":"string","description":"The uri for the children of a learning object","example":"https://api-gateway.clark.center/users/jdoe1/learning-objects/000000000000000000000000/children","required":false},"materials":{"type":"string","description":"The uri for the materials of a learning object","example":"https://api-gateway.clark.center/users/jdoe1/learning-objects/000000000000000000000000/materials","required":false},"metrics":{"type":"string","description":"The uri for the metrics of a learning object","example":"https://api-gateway.clark.center/users/jdoe1/learning-objects/000000000000000000000000/metrics","required":false},"parents":{"type":"string","description":"The uri for the parents of a learning object","example":"https://api-gateway.clark.center/users/jdoe1/learning-objects/000000000000000000000000/parents","required":false},"ratings":{"type":"string","description":"The uri for the ratings of a learning object","example":"https://api-gateway.clark.center/users/jdoe1/learning-objects/000000000000000000000000/ratings","required":false}}},"Changelog":{"properties":{"cuid":{"type":"string","description":"The cuid relating to the changelog","example":"00000000-1111-2222-3333-444444444444","required":true},"logs":{"type":"array","description":"The changes made to the object","required":true,"items":{"$ref":"#/components/schemas/ChangelogLog"}}}},"ChangelogLog":{"properties":{"date":{"type":"number","description":"The date the changelog was created in ms","example":0,"required":true},"text":{"type":"string","description":"The changelog text","example":"This is a change","required":true},"author":{"type":"object","$ref":"#/components/schemas/ChangelogAuthor"}}},"ChangelogAuthor":{"properties":{"userId":{"type":"string","description":"The id of the changelog author","example":0,"required":true},"name":{"type":"string","description":"The name of the changelog author","example":"John Doe","required":true},"role":{"type":"string","description":"The role of the changelog author","example":"admin","required":true},"profileImage":{"type":"string","description":"The profile image of the changelog author","required":true}}},"Metric":{"required":["collection","metrics"],"properties":{"collection":{"type":"string","description":"The collection pulling metrics from","example":"learning-objects"},"metrics":{"type":"object","description":"The metrics","$ref":"#/components/schemas/MetricBody"}}},"MetricBody":{"required":["saves","downloads"],"properties":{"saves":{"type":"integer","description":"The number of saves for a particular collection","example":200},"downloads":{"type":"integer","description":"The number of downloads for a particular collection","example":100},"topDownloads":{"type":"array","description":"The top downloads of a particular collection","items":{"$ref":"#/components/schemas/Download"}}}},"Download":{"required":["downloads","learningObjectCuid"],"properties":{"downloads":{"type":"integer","description":"The number of downloads for a particular learning object","example":100},"learningObjectCuid":{"type":"string","description":"The cuid of the learning object","example":"00000000-1111-2222-3333-444444444444"}}},"LibraryItem":{"properties":{"savedOn":{"type":"string","description":"When the object was saved to the user's library","example":0,"required":true},"savedBy":{"type":"string","description":"The username of the user who saved the object","example":"jdoe1","required":true},"learningObject":{"type":"object","description":"The learning object saved to the library","$ref":"#/components/schemas/LearningObject"}}},"Notification":{"properties":{"id":{"type":"string","description":"The id of the notification","example":0,"required":true},"text":{"type":"string","description":"The description of the notification","example":"This is an example","required":true},"timestamp":{"type":"string","description":"The timestamp for when the notification was created","example":0,"required":true},"link":{"type":"string","description":"The link to the object","example":"https://clark.center/details/jdoe1/00000000-1111-2222-3333-444444444444","required":true},"attributes":{"type":"object","required":true,"additionalProperties":{"type":"string"}}}},"Rating":{"properties":{"_id":{"type":"string","description":"The id of the rating object","example":0,"required":true},"value":{"type":"number","description":"The rating value, an integer, out of 5","example":4,"required":true},"user":{"type":"object","description":"The user who left the rating","$ref":"#/components/schemas/RatingUser","required":false},"comment":{"type":"string","description":"The comment the person made on the rating","example":"This is an example comment","required":true},"date":{"type":"number","description":"The date the person created the rating","example":0,"required":true},"response":{"type":"array","description":"An array of potential responses to the rating","items":{"$ref":"#/components/schemas/Response"},"required":false},"source":{"type":"object","description":"The learning object source information","$ref":"#/components/schemas/RatingSource","required":false}}},"Response":{"properties":{"_id":{"type":"string","description":"The id of the response object","example":0,"required":true},"user":{"type":"object","description":"The user who left the response","$ref":"#/components/schemas/RatingUser","required":true},"comment":{"type":"string","description":"The comment the person made on the response","example":"This is an example comment","required":true},"date":{"type":"number","description":"The date the person created the response","example":0,"required":true},"source":{"type":"object","description":"The learning object source information","$ref":"#/components/schemas/RatingSource"}}},"Flag":{"properties":{"_id":{"type":"string","description":"The id of the flag","example":0,"required":true},"comment":{"type":"string","description":"The comment for why the rating was flagged","example":"This is a flagged rating","required":true},"ratingId":{"type":"string","description":"The id of the rating flagged","example":0,"required":true},"date":{"type":"number","description":"The time when the rating was flagged","example":0,"required":true},"username":{"type":"string","description":"The username of the person who flagged the rating","example":"jdoe1","required":true},"concern":{"type":"string","description":"The type of the flag","example":"Other","required":true}}},"RatingUser":{"properties":{"username":{"type":"string","description":"The username of the person who left the rating","example":"jdoe1","required":true},"name":{"type":"string","description":"The name of the person who left the rating","example":"John Doe","required":true},"email":{"type":"string","description":"The email of the person who left the rating","example":"jdoe1@gmail.com","required":true}}},"RatingSource":{"properties":{"cuid":{"type":"string","description":"The cuid of the object","example":"00000000-1111-2222-3333-444444444444","required":true},"version":{"type":"number","description":"The version of the object","example":1,"required":true}}}}},"tags":[]}